Pull out Tracer from TestContainers.Docker module
